move fixtures to fixtures folder
